Safari-611.1.10/LayoutTests/fast/parser/iframe-sets-parent-to-javascript-url-expected.txt

ALERT: 1
ALERT: 2
ALERT: 3
ALERT: 4


--------
Frame: '<!--frame1-->'
--------
PASS
